requirements.txt
streamlit
python-dotenv
openai
google
google.genai
deep_translator